Extracting out rule loading so that rules can be stored in others other than on disk
