+++ elastalert/elastalert.py
@@ -17,11 +17,8 @@ from socket import error
 
 import dateutil.tz
 import kibana
-import yaml
 from alerts import DebugAlerter
-from config import get_rule_hashes
-from config import load_configuration
-from config import load_rules
+from config import load_conf
 from croniter import croniter
 from elasticsearch.exceptions import ConnectionError
 from elasticsearch.exceptions import ElasticsearchException
@@ -127,10 +124,10 @@ class ElastAlerter():
             tracer.setLevel(logging.INFO)
             tracer.addHandler(logging.FileHandler(self.args.es_debug_trace))
 
-        self.conf = load_rules(self.args)
+        self.conf = load_conf(self.args)
         self.max_query_size = self.conf['max_query_size']
         self.scroll_keepalive = self.conf['scroll_keepalive']
-        self.rules = self.conf['rules']
+        self.rules = self.conf['rules_loader'].load_all(self.conf, self.args)
         self.writeback_index = self.conf['writeback_index']
         self.run_every = self.conf['run_every']
         self.alert_time_limit = self.conf['alert_time_limit']
@@ -147,7 +144,7 @@ class ElastAlerter():
         self.current_es_addr = None
         self.buffer_time = self.conf['buffer_time']
         self.silence_cache = {}
-        self.rule_hashes = get_rule_hashes(self.conf, self.args.rule)
+        self.rule_hashes = self.conf['rules_loader'].get_hashes(self.conf, self.args)
         self.starttime = self.args.start
         self.disabled_rules = []
         self.replace_dots_in_field_names = self.conf.get('replace_dots_in_field_names', False)
@@ -967,9 +964,10 @@ class ElastAlerter():
         new_rule['filter'] = new_filters
 
     def load_rule_changes(self):
-        ''' Using the modification times of rule config files, syncs the running rules
-        to match the files in rules_folder by removing, adding or reloading rules. '''
-        new_rule_hashes = get_rule_hashes(self.conf, self.args.rule)
+        """ Using the modification times of rule config files, syncs the running rules
+            to match the files in rules_folder by removing, adding or reloading rules. """
+        rules_loader = self.conf['rules_loader']
+        new_rule_hashes = rules_loader.get_hashes(self.conf, self.args.rule)
 
         # Check each current rule for changes
         for rule_file, hash_value in self.rule_hashes.iteritems():
@@ -981,7 +979,7 @@ class ElastAlerter():
             if hash_value != new_rule_hashes[rule_file]:
                 # Rule file was changed, reload rule
                 try:
-                    new_rule = load_configuration(rule_file, self.conf)
+                    new_rule = rules_loader.load_configuration(rule_file, self.conf)
                     if 'is_enabled' in new_rule and not new_rule['is_enabled']:
                         elastalert_logger.info('Rule file %s is now disabled.' % (rule_file))
                         # Remove this rule if it's been disabled
@@ -991,12 +989,11 @@ class ElastAlerter():
                     message = 'Could not load rule %s: %s' % (rule_file, e)
                     self.handle_error(message)
                     # Want to send email to address specified in the rule. Try and load the YAML to find it.
-                    with open(rule_file) as f:
-                        try:
-                            rule_yaml = yaml.load(f)
-                        except yaml.scanner.ScannerError:
-                            self.send_notification_email(exception=e)
-                            continue
+                    try:
+                        rule_yaml = rules_loader.load_yaml(rule_file)
+                    except EAException:
+                        self.send_notification_email(exception=e)
+                        continue
 
                     self.send_notification_email(exception=e, rule=rule_yaml)
                     continue
@@ -1019,7 +1016,7 @@ class ElastAlerter():
         if not self.args.rule:
             for rule_file in set(new_rule_hashes.keys()) - set(self.rule_hashes.keys()):
                 try:
-                    new_rule = load_configuration(rule_file, self.conf)
+                    new_rule = rules_loader.load_configuration(rule_file, self.conf)
                     if 'is_enabled' in new_rule and not new_rule['is_enabled']:
                         continue
                     if new_rule['name'] in [rule['name'] for rule in self.rules]:
